+++ apps/api/src/app/benchmark/benchmark.module.ts
@@ -3,6 +3,7 @@ import { SymbolModule } from '@ghostfolio/api/app/symbol/symbol.module';
 import { ConfigurationModule } from '@ghostfolio/api/services/configuration/configuration.module';
 import { DataProviderModule } from '@ghostfolio/api/services/data-provider/data-provider.module';
 import { MarketDataModule } from '@ghostfolio/api/services/market-data/market-data.module';
+import { PrismaModule } from '@ghostfolio/api/services/prisma/prisma.module';
 import { PropertyModule } from '@ghostfolio/api/services/property/property.module';
 import { SymbolProfileModule } from '@ghostfolio/api/services/symbol-profile/symbol-profile.module';
 import { Module } from '@nestjs/common';
@@ -17,6 +18,7 @@ import { BenchmarkService } from './benchmark.service';
     ConfigurationModule,
     DataProviderModule,
     MarketDataModule,
+    PrismaModule,
     PropertyModule,
     RedisCacheModule,
     SymbolModule,